+++ tests/test_fileio.py
@@ -1,6 +1,6 @@
 import pytest
 
-from anyio import aopen
+from anyio import open_file
 
 pytestmark = pytest.mark.anyio
 
@@ -18,12 +18,12 @@ def testdatafile(tmp_path_factory, testdata):
 
 
 async def test_open_close(testdatafile):
-    f = await aopen(testdatafile)
-    await f.close()
+    f = await open_file(testdatafile)
+    await f.aclose()
 
 
 async def test_read(testdatafile, testdata):
-    async with await aopen(testdatafile, 'rb') as f:
+    async with await open_file(testdatafile, 'rb') as f:
         data = await f.read()
 
     assert f.closed
@@ -31,7 +31,7 @@ async def test_read(testdatafile, testdata):
 
 
 async def test_write(testdatafile, testdata):
-    async with await aopen(testdatafile, 'ab') as f:
+    async with await open_file(testdatafile, 'ab') as f:
         await f.write(b'f' * 1000)
 
     assert testdatafile.stat().st_size == len(testdata) + 1000
@@ -41,7 +41,7 @@ async def test_async_iteration(tmp_path):
     lines = ['blah blah\n', 'foo foo\n', 'bar bar']
     testpath = tmp_path.joinpath('testfile')
     testpath.write_text(''.join(lines), 'ascii')
-    async with await aopen(str(testpath)) as f:
+    async with await open_file(str(testpath)) as f:
         lines_i = iter(lines)
         async for line in f:
             assert line == next(lines_i)